[Refactor]: refactoring chat schecma
