+++ app/schemas/request/chat_request_schema.py
@@ -1,0 +1,7 @@
+from fastapi_camelcase import CamelModel
+
+
+class ChatCreate(CamelModel):
+    chat_id: int | None = None
+    user_id: int
+    character_id: int
