+++ app/schemas/response/chat_response_schema.py
@@ -1,0 +1,10 @@
+from fastapi_camelcase import CamelModel
+
+
+class ChatResponse(CamelModel):
+    chat_id: int
+    user_id: int
+    character_id: int
+
+class MessageResponse(CamelModel):
+    message: str
