+++ app/services/chat_service.py
@@ -3,7 +3,7 @@ from fastapi import HTTPException
 
 from ..crud import chat_crud
 from ..models.chat import Chat
-from ..schemas.chat_schema import ChatCreate
+from ..schemas.request.chat_request_schema import ChatCreate
 
 
 # insert
